Created Bus Factor, Correctness, Responsive Maintainer, License files that we can use to calculate the NetScore.
Was able to create separate files for Bus Factor, Correctness, Responsive Maintainer, License that we can use to figure out NetScore. Currently Bus Factor is the only file that calculates a value that we can use in NetScore. The rest of the files we will need to decide how to score the data numerically for the NetScore.