Updated CorrectnessMetric
Update the code to add a score out of 100. If the repo contain a README file then add 20 to the score. If the repo has a test file then add 50 to the score. Then if you have so many contributors it get a score based on that with a cap of 30 points for 30 contributors.